Builds as elf64 and boots to long mode. Interrupts not working and page tables hard coded.

+++ hw/isr.cpp
@@ -14,38 +14,38 @@ isr_t interruptHandlers[256];
  * of the function names */
 void isrInstall()
 {
-    setIDTGate(0, (UInt32)isr0);
-    setIDTGate(1, (UInt32)isr1);
-    setIDTGate(2, (UInt32)isr2);
-    setIDTGate(3, (UInt32)isr3);
-    setIDTGate(4, (UInt32)isr4);
-    setIDTGate(5, (UInt32)isr5);
-    setIDTGate(6, (UInt32)isr6);
-    setIDTGate(7, (UInt32)isr7);
-    setIDTGate(8, (UInt32)isr8);
-    setIDTGate(9, (UInt32)isr9);
-    setIDTGate(10, (UInt32)isr10);
-    setIDTGate(11, (UInt32)isr11);
-    setIDTGate(12, (UInt32)isr12);
-    setIDTGate(13, (UInt32)isr13);
-    setIDTGate(14, (UInt32)isr14);
-    setIDTGate(15, (UInt32)isr15);
-    setIDTGate(16, (UInt32)isr16);
-    setIDTGate(17, (UInt32)isr17);
-    setIDTGate(18, (UInt32)isr18);
-    setIDTGate(19, (UInt32)isr19);
-    setIDTGate(20, (UInt32)isr20);
-    setIDTGate(21, (UInt32)isr21);
-    setIDTGate(22, (UInt32)isr22);
-    setIDTGate(23, (UInt32)isr23);
-    setIDTGate(24, (UInt32)isr24);
-    setIDTGate(25, (UInt32)isr25);
-    setIDTGate(26, (UInt32)isr26);
-    setIDTGate(27, (UInt32)isr27);
-    setIDTGate(28, (UInt32)isr28);
-    setIDTGate(29, (UInt32)isr29);
-    setIDTGate(30, (UInt32)isr30);
-    setIDTGate(31, (UInt32)isr31);
+    setIDTGate(0, (UInt64)isr0);
+    setIDTGate(1, (UInt64)isr1);
+    setIDTGate(2, (UInt64)isr2);
+    setIDTGate(3, (UInt64)isr3);
+    setIDTGate(4, (UInt64)isr4);
+    setIDTGate(5, (UInt64)isr5);
+    setIDTGate(6, (UInt64)isr6);
+    setIDTGate(7, (UInt64)isr7);
+    setIDTGate(8, (UInt64)isr8);
+    setIDTGate(9, (UInt64)isr9);
+    setIDTGate(10, (UInt64)isr10);
+    setIDTGate(11, (UInt64)isr11);
+    setIDTGate(12, (UInt64)isr12);
+    setIDTGate(13, (UInt64)isr13);
+    setIDTGate(14, (UInt64)isr14);
+    setIDTGate(15, (UInt64)isr15);
+    setIDTGate(16, (UInt64)isr16);
+    setIDTGate(17, (UInt64)isr17);
+    setIDTGate(18, (UInt64)isr18);
+    setIDTGate(19, (UInt64)isr19);
+    setIDTGate(20, (UInt64)isr20);
+    setIDTGate(21, (UInt64)isr21);
+    setIDTGate(22, (UInt64)isr22);
+    setIDTGate(23, (UInt64)isr23);
+    setIDTGate(24, (UInt64)isr24);
+    setIDTGate(25, (UInt64)isr25);
+    setIDTGate(26, (UInt64)isr26);
+    setIDTGate(27, (UInt64)isr27);
+    setIDTGate(28, (UInt64)isr28);
+    setIDTGate(29, (UInt64)isr29);
+    setIDTGate(30, (UInt64)isr30);
+    setIDTGate(31, (UInt64)isr31);
 
     // Remap the PIC
     portByteOut(0x20, 0x11);
@@ -60,22 +60,22 @@ void isrInstall()
     portByteOut(0xA1, 0x0);
 
     // Install the IRQs
-    setIDTGate(32, (UInt32)irq0);
-    setIDTGate(33, (UInt32)irq1);
-    setIDTGate(34, (UInt32)irq2);
-    setIDTGate(35, (UInt32)irq3);
-    setIDTGate(36, (UInt32)irq4);
-    setIDTGate(37, (UInt32)irq5);
-    setIDTGate(38, (UInt32)irq6);
-    setIDTGate(39, (UInt32)irq7);
-    setIDTGate(40, (UInt32)irq8);
-    setIDTGate(41, (UInt32)irq9);
-    setIDTGate(42, (UInt32)irq10);
-    setIDTGate(43, (UInt32)irq11);
-    setIDTGate(44, (UInt32)irq12);
-    setIDTGate(45, (UInt32)irq13);
-    setIDTGate(46, (UInt32)irq14);
-    setIDTGate(47, (UInt32)irq15);
+    setIDTGate(32, (UInt64)irq0);
+    setIDTGate(33, (UInt64)irq1);
+    setIDTGate(34, (UInt64)irq2);
+    setIDTGate(35, (UInt64)irq3);
+    setIDTGate(36, (UInt64)irq4);
+    setIDTGate(37, (UInt64)irq5);
+    setIDTGate(38, (UInt64)irq6);
+    setIDTGate(39, (UInt64)irq7);
+    setIDTGate(40, (UInt64)irq8);
+    setIDTGate(41, (UInt64)irq9);
+    setIDTGate(42, (UInt64)irq10);
+    setIDTGate(43, (UInt64)irq11);
+    setIDTGate(44, (UInt64)irq12);
+    setIDTGate(45, (UInt64)irq13);
+    setIDTGate(46, (UInt64)irq14);
+    setIDTGate(47, (UInt64)irq15);
 
     setIDT(); // Load with ASM
 }
