+++ includes/kmem.h
@@ -26,7 +26,7 @@ public:
         UInt32 acpi30;
     } __attribute__((packed)) RegionInfo;
 
-    PhysicalMemoryManager(UInt32 memSize, UInt32 kernAddr, UInt32 kernSize,
+    PhysicalMemoryManager(UInt32 memSize, UInt64 kernAddr, UInt32 kernSize,
         RegionInfo *regions, UInt32 regionsLen);
     virtual ~PhysicalMemoryManager();
 
@@ -37,7 +37,7 @@ public:
     UInt32 memFreeBlocks();
     void *malloc(unsigned int size);
     void free(void *p);
-    void dropRegion(UInt32 base, UInt32 size); // allow drivers to reserve memory areas
+    void dropRegion(UInt64 base, UInt64 size); // allow drivers to reserve memory areas
 
 private:
     UInt32 physMemorySize;
@@ -75,7 +75,7 @@ private:
     inline int pmmBitmapTest(int bit);
     int findFirstFree();
     int findFirstFree(UInt32 size);
-    void initRegion(UInt32 base, UInt32 size);
+    void initRegion(UInt64 base, UInt64 size);
     void *allocBlock(); // alloc one block
     void *allocBlock(UInt32 size); // alloc size blocks
     void freeBlock(void *p); // free one block
